requirements.txt
hyundai_kia_connect_api>=3.24.0
beautifulsoup4>=4.10.0
paho_mqtt>=1.6.1
python_dateutil>=2.8.2
pytz>=2021.3
requests>=2.28.1
gspread>=5.6.2
